Moved oauth provider logos to assets folder
