+++ src/assets/logo_slack.png
