+++ src/pages/Welcome.js
@@ -22,9 +22,9 @@ import {
 } from '../styling/PageStyling';
 import { useForm } from 'react-hook-form';
 import { MdError } from 'react-icons/md';
-import Google from '../components/google-icon.svg';
-import Slack from '../components/logo_slack.png';
-import LinkedIn from '../components/logo_linkedin.png';
+import Google from '../assets/google-icon.svg';
+import Slack from '../assets/logo_slack.png';
+import LinkedIn from '../assets/logo_linkedin.png';
 
 const Welcome = () => {
